Better support for Policy OctoURI

+++ tests/test_policy_remote_e2e.py
@@ -7,6 +7,7 @@ import tempfile
 sys.path.append(".")
 
 from ghascompliance.policy import Policy
+from ghascompliance.utils.octouri import OctoUri
 
 
 class TestPolicyLoadingRemote(unittest.TestCase):
@@ -16,7 +17,7 @@ class TestPolicyLoadingRemote(unittest.TestCase):
 
     def tearDown(self):
 
-        if os.path.exists(self.policy.temp_repo):
+        if self.policy and os.path.exists(self.policy.temp_repo):
             shutil.rmtree(self.policy.temp_repo)
 
         return super().tearDown()
@@ -25,10 +26,12 @@ class TestPolicyLoadingRemote(unittest.TestCase):
         #  More of an integration test
         self.policy = Policy(
             severity="error",
-            repository="GeekMasher/advanced-security-compliance",
+            uri=OctoUri(
+                repository="GeekMasher/advanced-security-compliance",
+                path="examples/policies/test-policy.yml",
+                branch="testing",
+            ),
             token=os.environ.get("GITHUB_TOKEN"),
-            path="examples/policies/test-policy.yml",
-            branch="testing",
         )
 
         self.assertEqual(
